Add files via upload
This is the first (alpha) version of the filter list. I'm assuming it'll need more work, but let me know your what you think.
It's stored in a .txt file as a JavaScript array, where each element is a domain or part of one.
See https://adblockplus.org/filter-cheatsheet to see why (partial) domains look like that and see if we incorporate that.